Migrate to standard DDD structure (WIP)
