+++ cmd/productive-cli/main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"os"
 
-	"github.com/mister11/productive-cli/internal/app"
-	"github.com/mister11/productive-cli/internal/log"
+	"github.com/mister11/productive-cli/internal/infrastructure/log"
+	"github.com/mister11/productive-cli/internal/interfaces"
 )
 
 func main() {
 
-	if err := app.NewProductiveCLI().Run(os.Args); err != nil {
+	if err := interfaces.CLI().Run(os.Args); err != nil {
 		log.Error("CLI quit unexpectedly", err)
 		os.Exit(-1)
 	}
