--- internal/config/config_manager.go
@@ -1,11 +1,0 @@
-package config
-
-type ConfigManager interface {
-	GetToken() string
-	GetUserID() string
-	SaveToken(token string)
-	SaveUserID(userID string)
-	SaveProject(project Project)
-	GetSavedProjects() []Project
-	RemoveExistingProject(project Project)
-}
