+++ internal/domain/tracking/models.go
@@ -1,4 +1,4 @@
-package action
+package tracking
 
 type TrackFoodRequest struct {
 	IsWeekTracking bool
