+++ internal/infrastructure/log/log.go
@@ -2,8 +2,6 @@ package log
 
 import (
 	"fmt"
-	"os"
-
 	"github.com/fatih/color"
 )
 
@@ -15,6 +13,6 @@ func Info(format string, args ...interface{}) {
 	color.Magenta(format, args)
 }
 
-func Error(format string, err error) {
-	fmt.Fprintf(os.Stderr, format, err)
+func Error(format string, err ...error) {
+	_ = fmt.Errorf(format, err)
 }
