--- internal/stdin/promptui/models.go
@@ -1,7 +1,0 @@
-package promptui
-
-type PromptUiStdin struct{}
-
-func NewPromptUiStdin() *PromptUiStdin {
-	return &PromptUiStdin{}
-}
