+++ internal/utils/errors.go
@@ -1,7 +1,7 @@
 package utils
 
 import (
-	"github.com/mister11/productive-cli/internal/log"
+	"github.com/mister11/productive-cli/internal/infrastructure/log"
 )
 
 func ReportError(message string, err error) {
